revert jest, root-import and eslint-auto
