+++ .babelrc
@@ -21,9 +21,6 @@
       "displayName": true,
       "preprocess": false
     }],
-    "babel-plugin-inline-import-graphql-ast",
-    ["babel-plugin-root-import", {
-      "rootPathSuffix": "./"
-    }]
+    "babel-plugin-inline-import-graphql-ast"
   ]
 }
