--- components/LinkList/__tests__/__mocks__/index.mock.js
@@ -1,5 +1,0 @@
-export const defaultProps = {
-  pathname: 'signin',
-  authenticated: false,
-  logout: () => {}
-}
