--- components/LinkList/index.stories.js
@@ -1,6 +1,0 @@
-import React from 'react';
-
-import { storiesOf } from '@storybook/react';
-import LinkList from '.';
-
-storiesOf('Menu', module).add('LinkList', () => <LinkList />);
