+++ components/PostInfo/getPost.gql
