+++ components/PostUpvoter/upvotePost.gql
