+++ libraries/reducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import { reducer as authReducer } from './stores/auth';
+import { reducer as authReducer } from '../components/AuthFields/store';
 
 export default function getReducer(client) {
   return combineReducers({
