+++ package.json
@@ -33,15 +33,6 @@
       "git add"
     ]
   },
-  "jest": {
-    "moduleFileExtensions": [
-      "js"
-    ],
-    "testPathIgnorePatterns": [
-      "<rootDir>/(build|docs|node_modules)/",
-      "/__mocks__/"
-    ]
-  },
   "scripts": {
     "build": "next build",
     "dev": "npm run lint && node server.js",
@@ -61,13 +52,7 @@
     "release": "release",
     "graphql:play": "graphql playground",
     "graphql:update_schema": "graphql get-schema",
-    "graphql:see_graph": "graphql voyager",
-    "test": "NODE_ENV=test jest",
-    "test:watch": "npm run test -- --watch",
-    "test:update": "npm run test -- --u",
-    "test:debug": "node --inspect ./node_modules/.bin/jest --runInBand --env jest-environment-node-debug",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "graphql:see_graph": "graphql voyager"
   },
   "dependencies": {
     "babel-plugin-styled-components": "1.1.7",
@@ -100,18 +85,12 @@
   "devDependencies": {
     "babel-eslint": "7.2.3",
     "babel-plugin-inline-import-graphql-ast": "2.0.0",
-    "babel-plugin-root-import": "5.1.0",
     "cli-clear": "1.0.4",
-    "enzyme": "2.9.1",
-    "enzyme-to-json": "1.5.1",
     "eslint": "3.19.0",
     "eslint-config-airbnb": "15.1.0",
     "eslint-config-prettier": "2.3.0",
-    "eslint-import-resolver-babel-plugin-root-import": "https://github.com/bingqichen/eslint-import-resolver-babel-plugin-root-import",
-    "eslint-loader": "1.9.0",
     "eslint-plugin-graphql": "1.3.0",
     "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jest": "20.0.3",
     "eslint-plugin-jsx-a11y": "5.1.1",
     "eslint-plugin-prettier": "2.1.2",
     "eslint-plugin-react": "7.1.0",
@@ -122,18 +101,13 @@
     "handlebars": "4.0.10",
     "husky": "0.14.3",
     "inquirer": "3.2.1",
-    "jest": "20.0.4",
-    "jest-environment-node-debug": "2.0.0",
     "lint-staged": "4.0.3",
     "ngrok": "2.2.17",
     "nodemon": "1.11.0",
     "pre-commit": "1.2.2",
     "prettier": "1.5.3",
-    "react-addons-test-utils": "15.6.0",
-    "react-test-renderer": "15.6.1",
     "release": "1.3.3",
     "shelljs": "0.7.8",
-    "webpack-bundle-analyzer": "2.9.0",
-    "@storybook/react": "^3.2.3"
+    "webpack-bundle-analyzer": "2.9.0"
   }
 }
